Add www subdomain to nginx configuration
It exists for the LetsEncrypt section of the file, and without it
someone else might not realise they need to add it in order to get
LetsEncrypt working properly.

All of it is placeholders anyway, but the absence of the placeholder
doesn't make it explicitly obvious you have to add it, which could cause
someone to run into problems obliviously like I had.