Clarify configuration options for email set-up
This modifies the text explaining how emails are configured in
settings.py while also changing the placeholder values found in
env.default to make it more explicitly obvious what you need to write.

Otherwise someone might unknowingly be using an email address as a
username instead of the username attached to the account and are
subsequently assaulted by a bunch of errors which do nothing to indicate
this fact.

Not that I'm talking from personal experience.